requirements.txt
awswrangler
pandas
numpy
boto3
python_dateutil==2.8.2
reportlab==3.6.8
Requests==2.32.3
